requirements.txt
numpy=1.19.5
pandas=1.2.4
pytorch-lightning
pyyaml=5.3.1
scikit-learn=0.24.2
tables
tensorboard
tensorflow=2.5.0
